+++ src/components/ag-grid/plugins/utils.ts
@@ -1,0 +1,7 @@
+import { SortDirection } from "ag-grid-community";
+
+export function getNextSortOrder(current?: SortDirection): SortDirection {
+  if (!current) return "asc";
+  if (current === "asc") return "desc";
+  return null;
+}
