+++ src/modules/user/index.tsx
@@ -5,6 +5,5 @@ import { useCurrentUser } from "../../context/user.context";
 
 export default function User() {
   const { userData } = useCurrentUser();
-  console.log(userData);
   return <pre>{JSON.stringify(userData, null, 2)}</pre>;
 }
